cast expression added

+++ test.c
@@ -1,7 +1,7 @@
 void main(int x) {
 	int x = 5, b, c = 3;
 	float a;
-	a = x;
+	a = (float) x;
 
 	test("hello!");
 }
